refactor: separate express server setup
