+++ packages/call-service/src/app.ts
@@ -1,0 +1,16 @@
+import express from 'express';
+import cors from 'cors';
+import eventsRouter from './routes/events';
+import callsRouter from './routes/calls';
+
+const app = express();
+
+app.use(cors());
+app.use(express.json());
+
+app.use('/api/events', eventsRouter);
+app.use('/api/calls', callsRouter);
+
+app.get('/health', (_req, res) => res.json({ status: 'ok' }));
+
+export { app };
